Hide model objects, and do some more refactoring

+++ diff/diff.go
@@ -7,7 +7,9 @@ import (
 
 	"github.com/deckarep/golang-set"
 	"github.com/schemalex/schemalex"
-	"github.com/pkg/errors"
+	"github.com/schemalex/schemalex/format"
+	"github.com/schemalex/schemalex/internal/errors"
+	"github.com/schemalex/schemalex/model"
 )
 
 type Option interface {
@@ -40,21 +42,21 @@ func WithTransaction(b bool) Option {
 type diffCtx struct {
 	fromSet mapset.Set
 	toSet   mapset.Set
-	from    schemalex.Statements
-	to      schemalex.Statements
+	from    model.Stmts
+	to      model.Stmts
 }
 
-func newDiffCtx(from, to schemalex.Statements) *diffCtx {
+func newDiffCtx(from, to model.Stmts) *diffCtx {
 	fromSet := mapset.NewSet()
 	for _, stmt := range from {
-		if cs, ok := stmt.(*schemalex.CreateTableStatement); ok {
-			fromSet.Add(cs.Name)
+		if cs, ok := stmt.(model.Table); ok {
+			fromSet.Add(cs.ID())
 		}
 	}
 	toSet := mapset.NewSet()
 	for _, stmt := range to {
-		if cs, ok := stmt.(*schemalex.CreateTableStatement); ok {
-			toSet.Add(cs.Name)
+		if cs, ok := stmt.(model.Table); ok {
+			toSet.Add(cs.ID())
 		}
 	}
 
@@ -66,7 +68,7 @@ func newDiffCtx(from, to schemalex.Statements) *diffCtx {
 	}
 }
 
-func Statements(dst io.Writer, from, to schemalex.Statements, options ...Option) error {
+func Statements(dst io.Writer, from, to model.Stmts, options ...Option) error {
 	var txn bool
 	for _, o := range options {
 		switch o.Name() {
@@ -94,7 +96,7 @@ func Statements(dst io.Writer, from, to schemalex.Statements, options ...Option)
 		if err != nil {
 			return errors.Wrap(err, `failed to produce diff`)
 		}
-		if txn && n > 0 || !txn && buf.Len() > 0 && n > 0{
+		if txn && n > 0 || !txn && buf.Len() > 0 && n > 0 {
 			buf.WriteString("\n\n")
 		}
 		pbuf.WriteTo(&buf)
@@ -160,13 +162,23 @@ func Files(dst io.Writer, from, to string, options ...Option) error {
 
 func dropTables(ctx *diffCtx, dst io.Writer) (int64, error) {
 	var buf bytes.Buffer
-	names := ctx.fromSet.Difference(ctx.toSet)
-	for i, name := range names.ToSlice() {
+	ids := ctx.fromSet.Difference(ctx.toSet)
+	for i, id := range ids.ToSlice() {
 		if i > 0 {
 			buf.WriteByte('\n')
 		}
+
+		stmt, ok := ctx.from.Lookup(id.(string))
+		if !ok {
+			return 0, errors.Errorf(`failed to lookup table %s`, id)
+		}
+
+		table, ok := stmt.(model.Table)
+		if !ok {
+			return 0, errors.Errorf(`lookup failed: %s is not a model.Table`, id)
+		}
 		buf.WriteString("DROP TABLE `")
-		buf.WriteString(name.(string))
+		buf.WriteString(table.Name())
 		buf.WriteString("`;")
 	}
 
@@ -176,19 +188,19 @@ func dropTables(ctx *diffCtx, dst io.Writer) (int64, error) {
 func createTables(ctx *diffCtx, dst io.Writer) (int64, error) {
 	var buf bytes.Buffer
 
-	names := ctx.toSet.Difference(ctx.fromSet)
-	for _, name := range names.ToSlice() {
+	ids := ctx.toSet.Difference(ctx.fromSet)
+	for _, id := range ids.ToSlice() {
 		// Lookup the corresponding statement, and add its SQL
-		stmt, ok := ctx.to.Lookup(name.(string))
+		stmt, ok := ctx.to.Lookup(id.(string))
 		if !ok {
-			continue
+			return 0, errors.Errorf(`failed to lookup table %s`, id)
 		}
 
 		if buf.Len() > 0 {
 			buf.WriteByte('\n')
 		}
-		_, err := stmt.WriteTo(&buf)
-		if err != nil {
+
+		if err := format.SQL(&buf, stmt); err != nil {
 			return 0, err
 		}
 		buf.WriteByte(';')
@@ -201,29 +213,29 @@ type alterCtx struct {
 	toColumns   mapset.Set
 	fromIndexes mapset.Set
 	toIndexes   mapset.Set
-	from        *schemalex.CreateTableStatement
-	to          *schemalex.CreateTableStatement
+	from        model.Table
+	to          model.Table
 }
 
-func newAlterCtx(from, to *schemalex.CreateTableStatement) *alterCtx {
+func newAlterCtx(from, to model.Table) *alterCtx {
 	fromColumns := mapset.NewSet()
-	for _, col := range from.Columns {
-		fromColumns.Add(col.Name)
+	for col := range from.Columns() {
+		fromColumns.Add(col.ID())
 	}
 
 	toColumns := mapset.NewSet()
-	for _, col := range to.Columns {
-		toColumns.Add(col.Name)
+	for col := range to.Columns() {
+		toColumns.Add(col.ID())
 	}
 
 	fromIndexes := mapset.NewSet()
-	for _, idx := range from.Indexes {
-		fromIndexes.Add(idx.String())
+	for idx := range from.Indexes() {
+		fromIndexes.Add(idx.ID())
 	}
 
 	toIndexes := mapset.NewSet()
-	for _, idx := range to.Indexes {
-		toIndexes.Add(idx.String())
+	for idx := range to.Indexes() {
+		toIndexes.Add(idx.ID())
 	}
 
 	return &alterCtx{
@@ -245,23 +257,23 @@ func alterTables(ctx *diffCtx, dst io.Writer) (int64, error) {
 		addTableIndexes,
 	}
 
-	names := ctx.toSet.Intersect(ctx.fromSet)
+	ids := ctx.toSet.Intersect(ctx.fromSet)
 	var buf bytes.Buffer
-	for _, name := range names.ToSlice() {
-		var stmt schemalex.Stmt
+	for _, id := range ids.ToSlice() {
+		var stmt model.Stmt
 		var ok bool
 
-		stmt, ok = ctx.from.Lookup(name.(string))
+		stmt, ok = ctx.from.Lookup(id.(string))
 		if !ok {
-			return 0, errors.Errorf(`table '%s' not found in old schema (alter table)`, name)
+			return 0, errors.Errorf(`table '%s' not found in old schema (alter table)`, id)
 		}
-		beforeStmt := stmt.(*schemalex.CreateTableStatement)
+		beforeStmt := stmt.(model.Table)
 
-		stmt, ok = ctx.to.Lookup(name.(string))
+		stmt, ok = ctx.to.Lookup(id.(string))
 		if !ok {
-			return 0, errors.Errorf(`table '%s' not found in new schema (alter table)`, name)
+			return 0, errors.Errorf(`table '%s' not found in new schema (alter table)`, id)
 		}
-		afterStmt := stmt.(*schemalex.CreateTableStatement)
+		afterStmt := stmt.(model.Table)
 
 		var pbuf bytes.Buffer
 		alterCtx := newAlterCtx(beforeStmt, afterStmt)
@@ -290,9 +302,14 @@ func dropTableColumns(ctx *alterCtx, dst io.Writer) (int64, error) {
 			buf.WriteByte('\n')
 		}
 		buf.WriteString("ALTER TABLE `")
-		buf.WriteString(ctx.from.Name)
+		buf.WriteString(ctx.from.Name())
 		buf.WriteString("` DROP COLUMN `")
-		buf.WriteString(columnName.(string))
+		col, ok := ctx.from.LookupColumn(columnName.(string))
+		if !ok {
+			return 0, errors.Errorf(`failed to lookup column %s`, columnName)
+		}
+
+		buf.WriteString(col.Name())
 		buf.WriteString("`;")
 	}
 
@@ -313,9 +330,9 @@ func addTableColumns(ctx *alterCtx, dst io.Writer) (int64, error) {
 			buf.WriteByte('\n')
 		}
 		buf.WriteString("ALTER TABLE `")
-		buf.WriteString(ctx.from.Name)
+		buf.WriteString(ctx.from.Name())
 		buf.WriteString("` ADD COLUMN ")
-		if _, err := stmt.WriteTo(&buf); err != nil {
+		if err := format.SQL(&buf, stmt); err != nil {
 			return 0, err
 		}
 		buf.WriteByte(';')
@@ -346,11 +363,11 @@ func alterTableColumns(ctx *alterCtx, dst io.Writer) (int64, error) {
 			buf.WriteByte('\n')
 		}
 		buf.WriteString("ALTER TABLE `")
-		buf.WriteString(ctx.from.Name)
+		buf.WriteString(ctx.from.Name())
 		buf.WriteString("` CHANGE COLUMN `")
-		buf.WriteString(afterColumnStmt.Name)
+		buf.WriteString(afterColumnStmt.Name())
 		buf.WriteString("` ")
-		if _, err := afterColumnStmt.WriteTo(&buf); err != nil {
+		if err := format.SQL(&buf, afterColumnStmt); err != nil {
 			return 0, err
 		}
 		buf.WriteByte(';')
@@ -368,27 +385,27 @@ func dropTableIndexes(ctx *alterCtx, dst io.Writer) (int64, error) {
 			return 0, errors.Errorf(`index '%s' not found in old schema (drop index)`, index)
 		}
 
-		if indexStmt.Kind == schemalex.IndexKindPrimaryKey {
+		if indexStmt.IsPrimaryKey() {
 			if buf.Len() > 0 {
 				buf.WriteByte('\n')
 			}
 			buf.WriteString("ALTER TABLE `")
-			buf.WriteString(ctx.from.Name)
+			buf.WriteString(ctx.from.Name())
 			buf.WriteString("` DROP INDEX PRIMARY KEY;")
 			continue
 		}
 
-		if !indexStmt.Name.Valid {
-			return 0, errors.Errorf("can not drop index without name: %s", indexStmt.String())
+		if !indexStmt.HasName() {
+			return 0, errors.Errorf("can not drop index without name: %s", indexStmt.ID())
 		}
 
 		if buf.Len() > 0 {
 			buf.WriteByte('\n')
 		}
 		buf.WriteString("ALTER TABLE `")
-		buf.WriteString(ctx.from.Name)
+		buf.WriteString(ctx.from.Name())
 		buf.WriteString("` DROP INDEX `")
-		buf.WriteString(indexStmt.Name.Value)
+		buf.WriteString(indexStmt.Name())
 		buf.WriteString("`;")
 	}
 
@@ -407,9 +424,11 @@ func addTableIndexes(ctx *alterCtx, dst io.Writer) (int64, error) {
 			buf.WriteByte('\n')
 		}
 		buf.WriteString("ALTER TABLE `")
-		buf.WriteString(ctx.from.Name)
+		buf.WriteString(ctx.from.Name())
 		buf.WriteString("` ADD ")
-		buf.WriteString(indexStmt.String())
+		if err := format.SQL(&buf, indexStmt); err != nil {
+			return 0, err
+		}
 		buf.WriteByte(';')
 	}
